test: codify protected smoke contracts
